# swagger.yaml
openapi: 3.0.3
info:
  title: JetCV NFT API
  description: |
    API completa per la gestione di NFT tramite Crossmint e integrazione Veriff per autenticazione.

    ## 🎯 Funzionalità principali:
    - **NFT Management**: Mint, batch mint, aggiornamento metadati
    - **Wallet Management**: Creazione wallet Solana sicuri
    - **CV Management**: Validazione e creazione file CV con IPFS
    - **Collection Management**: Informazioni collezioni NFT
    - **Veriff Integration**: Autenticazione e verifica documenti

    ## 🔐 Autenticazione:
    - API Key per Crossmint
    - Chiavi pubbliche/private per Veriff

    ## 🌐 Ambienti:
    - Sviluppo: `http://localhost:4000`
    - Staging: `http://18.102.14.247:4000`
    - Produzione: `https://api.jetcv.com`
  version: 3.0.0
  contact:
    name: JetCV Development Team
    email: support@jetcv.com
    url: https://jetcv.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000
    description: Development server (local)
  - url: http://18.102.14.247:4000
    description: External server (staging)
  - url: https://api.jetcv.com
    description: Production server
...
paths:
...
  /api/veriff/session-request-veriff:
    post:
      tags: [Veriff]
      summary: Richiesta sessione Veriff
      description: Crea una sessione di verifica tramite l'API Veriff v1 con parametri personalizzabili
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                callback:
                  type: string
                  format: uri
                  example: "https://example.com/callback"
                  description: "URL di callback per le notifiche webhook (opzionale)"
                  default: "https://example.com/callback"
                firstName:
                  type: string
                  example: "riccardo"
                  description: "Nome della persona (opzionale)"
                  default: "riccardo"
                lastName:
                  type: string
                  example: "Doe"
                  description: "Cognome della persona (opzionale)"
                  default: "Doe"
                additionalFields:
                  type: object
                  description: "Campi aggiuntivi da includere nella richiesta (opzionale)"
                  example:
                    email: "riccardo.doe@example.com"
                    phoneNumber: "1234567890"
                    gender: "M"
                    dateOfBirth: "1990-01-01"
                    fullAddress: "123 Main Street, City, Country"
                    vendorData: "user-123"
                    endUserId: "riccardo-doe-123"
                  properties:
                    email:
                      {
                        type: string,
                        format: email,
                        description: "Indirizzo email",
                      }
                    phoneNumber:
                      { type: string, description: "Numero di telefono" }
                    gender:
                      {
                        type: string,
                        enum: ["M", "F"],
                        description: "Genere (M/F)",
                      }
                    dateOfBirth:
                      {
                        type: string,
                        format: date,
                        description: "Data di nascita (YYYY-MM-DD)",
                      }
                    fullAddress:
                      { type: string, description: "Indirizzo completo" }
                    vendorData:
                      {
                        type: string,
                        description: "Dati personalizzati del vendor",
                      }
                    endUserId: { type: string, description: "ID utente finale" }
      responses:
        "200":
          description: Sessione creata con successo
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message:
                    { type: string, example: "Test richiesta minima riuscito" }
                  receivedParams: { type: object }
                  builtData: { type: object }
                  response: { type: object }
                  sessionId:
                    {
                      type: string,
                      nullable: true,
                      example: "17e5c1b3-8b23-4a5a-8a09-f86bccaeb063",
                    }
                  sessionUrl:
                    {
                      type: string,
                      nullable: true,
                      example: "https://alchemy.veriff.com/session/17e5c1b3-8b23-4a5a-8a09-f86bccaeb063",
                    }
                  verificationUrl:
                    {
                      type: string,
                      nullable: true,
                      example: "https://station.veriff.com/sdk/17e5c1b3-8b23-4a5a-8a09-f86bccaeb063",
                    }
                  note:
                    {
                      type: string,
                      example: "Questa struttura dati funziona con l'API Veriff v1",
                    }
        "400":
          description: Errore nella richiesta
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  error:
                    { type: string, example: "Test richiesta minima fallito" }
                  details: { type: object }
                  status: { type: integer, example: 400 }
                  note:
                    {
                      type: string,
                      example: "Controlla i log per vedere l'errore specifico",
                    }
        "401":
          description: Autenticazione Veriff fallita
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  error: { type: string, example: "Errore API Veriff" }
                  details: { type: object }
                  status: { type: integer, example: 401 }
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  error: { type: string, example: "Errore interno del server" }
        "503":
          description: API Veriff non raggiungibile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  error:
                    {
                      type: string,
                      example: "Impossibile raggiungere l'API Veriff",
                    }
                  details:
                    { type: string, example: "Timeout o errore di connessione" }
...
  /api/veriff/session/{sessionId}:
    get:
      tags: [Veriff]
      summary: Verifica stato sessione
      description: Recupera lo stato di una sessione di verifica esistente tramite sessionId
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
          description: ID della sessione Veriff da verificare
          example: "17e5c1b3-8b23-4a5a-8a09-f86bccaeb063"
      responses:
        "200":
          description: Stato sessione recuperato con successo
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message:
                    {
                      type: string,
                      example: "Stato sessione recuperato con successo",
                    }
                  sessionId:
                    {
                      type: string,
                      example: "17e5c1b3-8b23-4a5a-8a09-f86bccaeb063",
                    }
                  sessionData:
                    {
                      type: object,
                      description: "Dati completi della sessione",
                    }
                  status:
                    {
                      type: string,
                      example: "approved",
                      description: "Status corrente della verifica",
                    }
                  createdAt:
                    {
                      type: string,
                      format: date-time,
                      description: "Data creazione sessione",
                    }
                  updatedAt:
                    {
                      type: string,
                      format: date-time,
                      description: "Data ultimo aggiornamento",
                    }
        "400":
          description: Errore nella richiesta
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  error: { type: string, example: "Errore recupero sessione" }
                  details: { type: object }
                  status: { type: integer, example: 400 }
                  sessionId: { type: string }
        "401":
          description: Autenticazione Veriff fallita
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  error: { type: string, example: "Errore API Veriff" }
                  details: { type: object }
                  status: { type: integer, example: 401 }
        "404":
          description: Sessione non trovata
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  error: { type: string, example: "Sessione non trovata" }
                  details: { type: object }
                  status: { type: integer, example: 404 }
  /api/veriff/verification/{verificationId}:
    get:
      tags: [Veriff]
      summary: Dettagli verifica completata
      description: Recupera i dettagli completi di una verifica completata tramite verificationId
      parameters:
        - name: verificationId
          in: path
          required: true
          schema:
            type: string
          description: ID della verifica completata
          example: "ver_1234567890abcdef"
      responses:
        "200":
          description: Dettagli verifica recuperati con successo
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message:
                    {
                      type: string,
                      example: "Dettagli verifica recuperati con successo",
                    }
                  verificationId:
                    { type: string, example: "ver_1234567890abcdef" }
                  verificationData:
                    {
                      type: object,
                      description: "Dati completi della verifica",
                    }
                  status:
                    {
                      type: string,
                      example: "approved",
                      description: "Status finale della verifica",
                    }
                  person:
                    {
                      type: object,
                      description: "Dati della persona verificata",
                    }
                  document:
                    {
                      type: object,
                      description: "Dati del documento verificato",
                    }
                  address: { type: object, description: "Indirizzo verificato" }
                  createdAt:
                    {
                      type: string,
                      format: date-time,
                      description: "Data creazione verifica",
                    }
                  updatedAt:
                    {
                      type: string,
                      format: date-time,
                      description: "Data ultimo aggiornamento",
                    }
        "400":
          description: Errore nella richiesta
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  error: { type: string, example: "Errore recupero verifica" }
                  details: { type: object }
                  status: { type: integer, example: 400 }
                  verificationId: { type: string }
        "401":
          description: Autenticazione Veriff fallita
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  error: { type: string, example: "Errore API Veriff" }
                  details: { type: object }
                  status: { type: integer, example: 401 }
        "404":
          description: Verifica non trovata
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  error: { type: string, example: "Verifica non trovata" }
                  details: { type: object }
                  status: { type: integer, example: 404 }
  /api/veriff/webhook:
    post:
      tags: [Veriff]
      summary: Webhook Veriff
      description: Endpoint per ricevere notifiche webhook da Veriff sui cambiamenti di stato delle verifiche
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Dati del webhook inviati da Veriff"
              properties:
                type:
                  type: string
                  description: "Tipo di evento webhook"
                  example: "verification_approved"
                  enum:
                    [
                      "session_created",
                      "verification_approved",
                      "verification_declined",
                      "verification_expired",
                    ]
                verification:
                  type: object
                  description: "Dati della verifica"
                  properties:
                    id:
                      {
                        type: string,
                        description: "ID della sessione/verifica",
                      }
                    status:
                      {
                        type: string,
                        description: "Nuovo status della verifica",
                      }
                    person: { type: object, description: "Dati della persona" }
                    document:
                      { type: object, description: "Dati del documento" }
                    address:
                      { type: object, description: "Indirizzo verificato" }
                    createdAt: { type: string, format: date-time }
                    updatedAt: { type: string, format: date-time }
      responses:
        "200":
          description: Webhook processato con successo
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message:
                    { type: string, example: "Webhook processato con successo" }
                  eventType: { type: string, example: "verification_approved" }
                  sessionId:
                    {
                      type: string,
                      example: "17e5c1b3-8b23-4a5a-8a09-f86bccaeb063",
                    }
                  status: { type: string, example: "approved" }
                  timestamp: { type: string, format: date-time }
        "401":
          description: Firma webhook non valida
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string, example: "Firma non valida" }
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  error:
                    { type: string, example: "Errore processamento webhook" }
                  details: { type: string }
  /api/veriff/test:
    get:
      tags: [Veriff]
      summary: Test configurazione Veriff
      description: Endpoint di test per verificare la configurazione delle API Veriff
      responses:
        "200":
          description: Test configurazione riuscito
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: "API Veriff funzionante" }
                  config:
                    type: object
                    properties:
                      publicKey: { type: string, example: "Configurato" }
                      privateKey: { type: string, example: "Configurato" }
                      baseUrl:
                        {
                          type: string,
                          example: "https://stationapi.veriff.com/v1",
                        }
                  endpoints:
                    type: object
                    properties:
                      createSession:
                        {
                          type: string,
                          example: "POST /api/veriff/session-request-veriff",
                        }
                      getSession:
                        {
                          type: string,
                          example: "GET /api/veriff/session/:sessionId",
                        }
                      getVerification:
                        {
                          type: string,
                          example: "GET /api/veriff/verification/:verificationId",
                        }
                      webhook:
                        { type: string, example: "POST /api/veriff/webhook" }
                      test: { type: string, example: "GET /api/veriff/test" }
                  timestamp: { type: string, format: date-time }
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  error: { type: string, example: "Errore test API" }
                  details: { type: string }
...
security:
  - ApiKeyAuth: []